+++ CHANGES.md
@@ -111,6 +111,12 @@
       XMobar so that padding space may be reserved on xmobar for the
       tray.  Requires `xmobar` version 0.40 or higher.
 
+- `XMonad.Layout.VoidBorders`
+
+    - Added new layout modifier `normalBorders` which can be used for
+      resetting borders back in layouts where you want borders after calling
+      `voidBorders`.
+
 ## 0.17.0 (October 27, 2021)
 
 ### Breaking Changes